# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Metaphysics Engine API",
    "version": "2.0.0",
    "description": "中国传统术数的算法能力层：八字、紫微斗数、六爻纳甲、大六壬、奇门遁甲、八宅风水、\n择吉、姓名五格，以及塔罗、周易起卦、星座、合盘。\n\n**结构归引擎，断语归调用方**：盘怎么排（三传、安星、排局、游年）有唯一正确答案，\n引擎必须算准；盘怎么解（庙旺利陷、奇门格局）各家分歧大，引擎只把断语所需的原料给全。\n有流派分歧的排盘口径（藏干权重、闰月归属、拆补定局、转盘排星）在各端点的 description 里注明。\n\n**无状态**：不存数据、不认用户、没有数据库。同样的入参永远得到同样的结果，\n可以随意水平扩容，也不需要迁移或备份。\n\n**无鉴权**：所有端点都是公开的，靠限流而不是身份来控制成本。\n需要访问控制请放在反向代理层。唯一带凭据的是 `/api-docs`（Basic）和 `/metrics`（Bearer），\n它们都不是能力端点。\n\n`/ai-interpret` 这类端点会调用外部大模型，是唯一有副作用（花钱、可能超时）的一类；\n排盘类端点纯本地计算。"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "API 服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "metricsToken": {
        "type": "http",
        "scheme": "bearer"
      },
      "docsBasic": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "错误信息"
          }
        }
      },
...
      "Pillars": {
        "type": "object",
        "properties": {
          "year": {
            "$ref": "#/components/schemas/Pillar"
          },
          "month": {
            "$ref": "#/components/schemas/Pillar"
          },
          "day": {
            "$ref": "#/components/schemas/Pillar"
          },
          "hour": {
            "$ref": "#/components/schemas/Pillar"
          }
        }
      },
...
      "BaziCalculationRequest": {
        "type": "object",
        "required": [
          "birthYear",
          "birthMonth",
          "birthDay",
          "birthHour",
          "gender"
        ],
        "properties": {
          "birthYear": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9999,
            "example": 1990
          },
          "birthMonth": {
            "type": "integer",
            "minimum": 1,
            "maximum": 12,
            "example": 5
          },
          "birthDay": {
            "type": "integer",
            "minimum": 1,
            "maximum": 31,
            "example": 20
          },
          "birthHour": {
            "type": "integer",
            "minimum": 0,
            "maximum": 23,
            "example": 14
          },
          "birthMinute": {
            "type": "integer",
            "minimum": 0,
            "maximum": 59,
            "default": 0
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female"
            ],
            "example": "male"
          },
          "birthLocation": {
            "type": "string",
            "description": "地名或 \"纬度,经度\" 坐标串。认得的地名见 GET /api/locations；解析不出经度时不做真太阳时校正，但排盘照常返回。",
            "example": "Beijing"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Shanghai"
          },
          "timezoneOffsetMinutes": {
            "type": "integer",
            "description": "给了就优先于 timezone，用于无法解析 IANA 时区名的调用方。"
          }
        }
      },
      "BaziCalculation": {
        "type": "object",
        "properties": {
          "pillars": {
            "$ref": "#/components/schemas/Pillars"
          },
          "fiveElements": {
            "type": "object",
            "description": "五行计数。",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "fiveElementsPercent": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "tenGods": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "strength": {
                  "type": "number"
                }
              }
            }
          },
          "luckCycles": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "八步大运，每步带十神、纳音与逐年流年"
          },
          "luckStart": {
            "type": "object",
            "description": "起运还需几年几月几天，以及交运的公历日期"
          },
          "analysis": {
            "type": "object",
            "description": "断命层，也是该拿去做判断的那一份：藏干加权五行、身强身弱、扶抑法用神喜忌、逐柱藏干与十神、神煞、四柱刑冲合会、旬空。与顶层 fiveElements 的个数统计不是一回事。"
          },
          "chartTime": {
            "type": "object",
            "description": "实际用于排盘的时刻。真太阳时生效时 used 与 trueSolarTime.clockTime 会不同，差一个时辰即差一柱。locationResolution 说明出生地解析的结果。",
            "properties": {
              "used": {
                "type": "object"
              },
              "trueSolarTime": {
                "$ref": "#/components/schemas/TrueSolarTime"
              },
              "locationResolution": {
                "$ref": "#/components/schemas/LocationResolution"
              }
            }
          },
          "strength": {
            "type": "object"
          },
          "timezoneOffsetMinutes": {
            "type": "integer",
            "nullable": true
          },
          "trueSolarTime": {
            "$ref": "#/components/schemas/TrueSolarTime"
          },
          "locationResolution": {
            "$ref": "#/components/schemas/LocationResolution"
          }
        }
      },
      "AiContent": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "Markdown 文本"
          }
        }
      },
      "AiInterpretRequestBase": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "description": "覆盖默认供应商。可用值见 GET /api/ai/providers；不可用会返回 400。",
            "enum": [
              "openai",
              "anthropic",
              "mock"
            ]
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/bazi/calculate": {
      "post": {
        "tags": [
          "八字"
        ],
        "summary": "八字排盘",
        "description": "纯计算。命中缓存时响应头 `x-bazi-cache: hit`，否则 `miss`——缓存只影响延迟，不影响结果。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaziCalculationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaziCalculation"
                }
              }
            }
          },
          "400": {
            "description": "入参非法（含非法日期、纯空白字符串）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "计算失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/bazi/ai-interpret": {
      "post": {
        "tags": [
          "八字",
          "AI"
        ],
        "summary": "八字 AI 解读",
        "description": "接收 /api/bazi/calculate 的输出，返回解读文本。同一调用方同时只允许一个 AI 请求。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AiInterpretRequestBase"
                  },
                  {
                    "type": "object",
                    "required": [
                      "pillars"
                    ],
                    "properties": {
                      "pillars": {
                        "$ref": "#/components/schemas/Pillars"
                      },
                      "fiveElements": {
                        "type": "object"
                      },
                      "tenGods": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        }
                      },
                      "luckCycles": {
                        "type": "array",
                        "items": {
                          "type": "object"
                        }
                      },
                      "strength": {
                        "type": "object"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI 解读文本（Markdown）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiContent"
                }
              }
            }
          },
          "400": {
            "description": "入参缺失，或指定了不可用的 provider",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "同一调用方已有一个 AI 请求在处理中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "上游 AI 供应商不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/bazi/full-analysis": {
      "post": {
        "tags": [
          "八字",
          "AI"
        ],
        "summary": "八字排盘 + AI 解读（一次调用）",
        "description": "calculate 与 ai-interpret 的合并调用，省一个来回。入参同 calculate。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AiInterpretRequestBase"
                  },
                  {
                    "$ref": "#/components/schemas/BaziCalculationRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaziCalculation"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "calculation": {
                          "$ref": "#/components/schemas/BaziCalculation"
                        },
                        "interpretation": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "入参非法，或指定了不可用的 provider",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "同一调用方已有一个 AI 请求在处理中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "分析失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tarot/ai-interpret": {
      "post": {
        "tags": [
          "塔罗",
          "AI"
        ],
        "summary": "塔罗 AI 解读",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AiInterpretRequestBase"
                  },
                  {
                    "type": "object",
                    "required": [
                      "cards"
                    ],
                    "properties": {
                      "spreadType": {
                        "type": "string"
                      },
                      "cards": {
                        "type": "array",
                        "minItems": 1,
                        "items": {
                          "$ref": "#/components/schemas/TarotCard"
                        }
                      },
                      "userQuestion": {
                        "type": "string"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI 解读文本（Markdown）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiContent"
                }
              }
            }
          },
          "400": {
            "description": "入参缺失，或指定了不可用的 provider",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "同一调用方已有一个 AI 请求在处理中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "上游 AI 供应商不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/iching/ai-interpret": {
      "post": {
        "tags": [
          "周易",
          "AI"
        ],
        "summary": "周易 AI 解读",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AiInterpretRequestBase"
                  },
                  {
                    "type": "object",
                    "required": [
                      "hexagram"
                    ],
                    "properties": {
                      "hexagram": {
                        "oneOf": [
                          {
                            "type": "string"
                          },
                          {
                            "$ref": "#/components/schemas/Hexagram"
                          }
                        ],
                        "description": "卦名字符串或完整卦象对象都接受。"
                      },
                      "userQuestion": {
                        "type": "string"
                      },
                      "method": {
                        "type": "string"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI 解读文本（Markdown）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiContent"
                }
              }
            }
          },
          "400": {
            "description": "入参缺失，或指定了不可用的 provider",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "同一调用方已有一个 AI 请求在处理中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "上游 AI 供应商不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}